Add CUDA support for BoTorch (#19)
Allow specifying the compute device for the PyTorch objects in BoTorch.
This allows running the optimisation in the GPU with CUDA which,
locally, has been observed to be almost 6x faster than the CPU version.

However, GPU memory may be a limitation. Thus, when a CUDA out-of-memory
error is raised, the optimisation falls back to the CPU automatically.

While there, fix an issue with standardising the test dataset in
BoTorch.